+++ Changelog.md
@@ -1,3 +1,11 @@
+### dev
+[full changelog](http://github.com/rspec/rspec-rails/compare/v2.13.2...master)
+
+Enhancements
+
+* Prelimiarily support Rails 4.1 by updating adapters to support Minitest 5.0.
+  (Andy Lindeman)
+
 ### 2.13.2 / 2013-05-18
 [full changelog](http://github.com/rspec/rspec-rails/compare/v2.13.1...v2.13.2)